♻️refactor: unify solicitations controller
