--- src/main/java/br/com/chronos/server/api/controllers/solicitation/time_punch_log_adjustments/TimePunchLogAdjustmentController.java
@@ -1,17 +1,0 @@
-package br.com.chronos.server.api.controllers.solicitation.time_punch_log_adjustments;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@RestController
-@RequestMapping("/solicitation/time-punch-adjustment")
-public @interface TimePunchLogAdjustmentController {
-
-}
